Magpie 5 and More Transformations
Put the noncommittal into an array for convenience.
Transformed “is” statement into a “Why is” question.
Transformed “are” statement into a “Why are” question.
Transformed “was” statement into a “Why was” question.
Transformed “You are” statement into a “Why am I” question
Replies with “Why do you say that?” to a “I said” statement.